added registry model protos and cleaned up strucutre
