+++ internal/motor/proto/registry/v1/account.proto
@@ -1,0 +1,16 @@
+syntax = "proto3";
+
+package sonrio.motor.registry.v1;
+option go_package = "github.com/sonr-io/sonr/motor/types/registry";
+
+// CreateAccount Request contains the three keys needed to create an account on Sonr
+message CreateAccountRequest {
+  string password = 1;
+  bytes ecdsa_dsc_key = 2;
+  bytes ecdsa_preshared_key = 3;
+}
+
+message CreateAccountResponse {
+  bytes psk = 1;
+  string address = 2;
+}
